ml-service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML & Deep Learning
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2  # ADD THIS LINE

# NLP & Text Processing
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.37.2

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2